+++ test/component/cypress/specs/client/modules/tasks/TaskDetails/TaskButtons.cy.jsx
@@ -1,0 +1,42 @@
+import React from 'react'
+
+import { taskWithInvestmentProjectFaker } from '../../../../../../../functional/cypress/fakers/task'
+import TaskDetailsTable from '../../../../../../../../src/client/modules/Tasks/TaskDetails/TaskDetailsTable'
+import DataHubProvider from '../../../../provider'
+
+describe('Task buttons', () => {
+  const Component = (props) => (
+    <DataHubProvider>
+      <TaskDetailsTable {...props} />
+    </DataHubProvider>
+  )
+
+  context('When a task is not completed', () => {
+    const investmentProjectTask = taskWithInvestmentProjectFaker()
+    const company =
+      investmentProjectTask.investmentProjectTask.investmentProject
+        .investorCompany
+
+    it('should show the Mark as complete button', () => {
+      cy.mount(<Component task={investmentProjectTask} company={company} />)
+      cy.get('[data-test="submit-button"]').should(
+        'contain.text',
+        'Mark as complete'
+      )
+    })
+  })
+
+  context('When a task is completed', () => {
+    const investmentProjectTask = taskWithInvestmentProjectFaker({
+      archived: true,
+    })
+    const company =
+      investmentProjectTask.investmentProjectTask.investmentProject
+        .investorCompany
+
+    it('should not show the Mark as complete button', () => {
+      cy.mount(<Component task={investmentProjectTask} company={company} />)
+      cy.get('[data-test="submit-button"]').should('not.exist')
+    })
+  })
+})
