+++ test/component/cypress/specs/client/modules/tasks/TaskDetails/TaskDetailsTable.cy.jsx
@@ -1,17 +1,19 @@
 import React from 'react'
 
-import {
-  assertFormButtons,
-  assertSummaryTable,
-} from '../../../../../../../functional/cypress/support/assertions'
+import { assertSummaryTable } from '../../../../../../../functional/cypress/support/assertions'
 import { taskWithInvestmentProjectFaker } from '../../../../../../../functional/cypress/fakers/task'
 import urls from '../../../../../../../../src/lib/urls'
 import { formatLongDate } from '../../../../../../../../src/client/utils/date'
 import { NOT_SET_TEXT } from '../../../../../../../../src/apps/companies/constants'
 import TaskDetailsTable from '../../../../../../../../src/client/modules/Tasks/TaskDetails/TaskDetailsTable'
+import DataHubProvider from '../../../../provider'
 
-describe('ProjectLayout', () => {
-  const Component = (props) => <TaskDetailsTable {...props} />
+describe('Task details table', () => {
+  const Component = (props) => (
+    <DataHubProvider>
+      <TaskDetailsTable {...props} />
+    </DataHubProvider>
+  )
 
   context('When a task has all optional fields set', () => {
     const investmentProjectTask = taskWithInvestmentProjectFaker()
@@ -76,35 +78,6 @@ describe('ProjectLayout', () => {
           'Created by': investmentProjectTask.createdBy.name,
         },
       })
-    })
-  })
-
-  context('When a task is not completed', () => {
-    const investmentProjectTask = taskWithInvestmentProjectFaker()
-    const company =
-      investmentProjectTask.investmentProjectTask.investmentProject
-        .investorCompany
-
-    it('should show the Mark as complete button', () => {
-      cy.mount(<Component task={investmentProjectTask} company={company} />)
-      assertFormButtons(
-        'Mark as complete',
-        'cancel',
-        urls.investments.projects.tasks.index(
-          investmentProjectTask.investmentProject.id
-        )
-      )
-    })
-  })
-
-  context('When a task is completed', () => {
-    const investmentProjectTask = taskWithInvestmentProjectFaker()
-    const company =
-      investmentProjectTask.investmentProjectTask.investmentProject
-        .investorCompany
-
-    it('should not show the Mark as complete button', () => {
-      cy.mount(<Component task={investmentProjectTask} company={company} />)
     })
   })
 })
