+++ test/functional/cypress/specs/tasks/details-spec.js
@@ -1,6 +1,12 @@
 import { tasks, dashboard, companies } from '../../../../../src/lib/urls'
 import { taskWithInvestmentProjectFaker } from '../../fakers/task'
-import { assertSummaryTable, assertBreadcrumbs } from '../../support/assertions'
+import { clickButton } from '../../support/actions'
+import {
+  assertSummaryTable,
+  assertBreadcrumbs,
+  assertPayload,
+  assertUrl,
+} from '../../support/assertions'
 
 describe('View task details', () => {
   const investmentProjectTask = taskWithInvestmentProjectFaker()
@@ -38,6 +44,19 @@ describe('View task details', () => {
         assertSummaryTable({
           dataTest: 'task-details-table',
         })
+      })
+
+      it('should redirect to the investment project and show the Flash message after marking as complete', () => {
+        cy.intercept(
+          'POST',
+          `/api-proxy/v4/task/${investmentProjectTask.id}/archive`,
+          {}
+        ).as('postTaskArchiveApiRequest')
+        clickButton('Mark as complete')
+        assertPayload('@postTaskArchiveApiRequest', { reason: 'completed' })
+        assertUrl(
+          `/investments/projects/${investmentProjectTask.investmentProjectTask.investmentProject.id}/tasks`
+        )
       })
     }
   )
